fix(ci): pin the toolchain — unpinned linter turned a green tree red
CI had been failing on every push while local runs were green. Root cause: the
workflow ran `pip install ruff mypy …` unpinned, so CI silently upgraded to ruff
0.16.0, whose new rules flagged 50 pre-existing lines. A green tree can go red with
no code change — the definition of a non-reproducible build, and a direct violation
of Forge's own ENGINEERING.md rule ("MUST: pin dependencies"). Applied that rule to
Forge itself: ruff==0.15.17, mypy==2.1.0.

Upgrading the linter is now a deliberate, separate task rather than something CI does
implicitly on a random Tuesday.

Also in this commit (found while diagnosing):
- forge_calibrate.py: adaptive calibration — reads the project's own run history and
  reports which limits are miscalibrated TOO TIGHT (fires on honest work -> users
  disable it) or TOO LOOSE (never fires -> security theatre), with the evidence and
  the exact command to apply. It never edits config: a control that loosens itself is
  a control that disarms itself, so Forge proposes and the human decides.
- Restored `# noqa: E402` on status/forge_dashboard.py's sys.path-dependent imports,
  stripped by an autofix run under the newer ruff.
- Fixed a latent runtime bug I introduced mid-fix (pathlib used without import in
  render_dashboard.py — syntax-valid, would have crashed the CLI path; tests didn't
  cover it).

Gate verified locally with the exact pinned versions CI uses: ruff clean, mypy clean,
324 tests, 91% coverage, guard proof 100%/0%, self-audit OK 10/10.

+++ traces/render_pipeline.py
@@ -97,7 +97,7 @@ def _parse_ts(ev: dict[str, object]) -> datetime | None:
 
 
 def fmt_duration(seconds: float) -> str:
-    total = int(round(seconds))
+    total = round(seconds)
     if total <= 0:
         return "0s"
     minutes, secs = divmod(total, 60)
